Criado planos e detalhes do plano e permissoes

+++ app/Http/Controllers/Admin/PlanController.php
@@ -81,10 +81,19 @@ class PlanController extends Controller
      */
     public function destory($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where('url', $url)
+            ->first();
 
         if (empty($plan))
             return redirect()->back();
+
+        if ($plan->details->count() > 0){
+            return redirect()
+                ->back()
+                ->with('error', 'Existem detalhes vinculados a este plano, portanto é necessário excluir os detalhes antes de deletar o plano');
+        }
 
         $plan->delete();
